Various fixes
closes #4  closes #3  closes #2  closes #1  closes #7

+++ main.cpp
@@ -10,12 +10,9 @@ int main()
 
     Network local_network;
 
-    // local_network.import_network("test_network.txt");
-    //system("CLS");
-
     bool loop1 = true, gen = false;
     short selec;
-    int cost, key;
+    int cost;
 
     std::string menu1 =
             "1. Importar red\n"
@@ -23,7 +20,7 @@ int main()
             "3. Crear red manualmente\n"
             "4. Salir\n"
 
-            , name1, name2,
+            , name1, name2, aux,
 
             menu2 =
             "1. Agregar router\n"
@@ -35,17 +32,14 @@ int main()
             "7. Exportar red\n"
             "8. Ver tabla de enrutamiento\n"
             "9. Encontrar camino mas eficiente entre 2 routers\n"
-            "10. Crear nueva red\n";
-
+            "10. Crear nueva red\n"
+            "11. Salir\n";
 
     //Menu starts here
 
     while(loop1){
 
-        //something that checks is the network is empty
-        // and if all nodes have at least one link****************
-
-        //make something to get out of menus
+        //make something to get out of menus***
 
         if(gen == false){
 
@@ -60,24 +54,28 @@ int main()
                 if(name1 != "-1"){
                     system("CLS");
                     local_network.import_network(name1);
-                    gen = true;
+
+                    if(local_network.is_empty() == false)
+                        gen = true;
                 }
                 break;
             }
             case 2:{ //Generar red
                 local_network.generate_network();
                 gen = true;
+                std::cout<<"Red generada correctamente"<<std::endl;
+                Sleep(1000);
                 break;
             }
             case 3:{ //Crear red manualmente
 
                 std::cout<<"Para crear una red debes agregar minimo 2 routers y un enlace ellos. Podras modificar la red despues de crearla"<<std::endl;
 
-                std::cout<<"Ingresa el nombre del primer router: "<<std::endl;
+                std::cout<<"Ingresa el nombre del primer router: ";
                 std::cin>>name1;
                 local_network.add_router(name1);
 
-                std::cout<<"Ingresa el nombre del segundo router: "<<std::endl;
+                std::cout<<"Ingresa el nombre del segundo router: ";
                 std::cin>>name2;
                 local_network.add_router(name2);
 
@@ -99,7 +97,10 @@ int main()
             }
         }
 
-        if(gen == true){ // if there's an existing network
+        local_network.verify_integrity();               // checks the integrity of the network and each router.
+        if(local_network.is_empty()) gen = false; // Removes routers that are left alone, and takes the user back to the network creation menu when the network is not valid anymore
+
+        else if(gen == true){ // if there's an existing network
 
             selec = menu_selection(menu2);
 
@@ -152,7 +153,25 @@ int main()
                     }
                 }
 
-                local_network.delete_router(name1);
+                while(true){ //confirmation
+
+                    std::cout<<"Esta a punto de borrar el router "<<name1<<" y todos sus enlaces (Esto puede afectar la existencia de otros routers). Desea continuar? (Y/N): "; std::cin>>name2;
+
+                    if(name2 == "Y" || name2 == "y"){
+
+                        local_network.delete_router(name1);
+
+                        std::cout<<"Router borrado correctamente";
+                        Sleep(1000);
+
+                        break;
+                    }
+                    else if(name2 == "N" || name2 == "n") break;
+                    else {
+                        system("CLS");
+                        std::cout<<"Opcion no valida"<<std::endl;
+                    }
+                }
 
                 break;
             }
@@ -183,26 +202,35 @@ int main()
             case 4:{ //Agregar enlace
 
                 while(true){
-                    std::cout<<"Ingrese el router origen: ";
-                    std::cin>>name1;
+                    while(true){
+                        std::cout<<"Ingrese el router origen: ";
+                        std::cin>>name1;
 
-                    if(local_network.is_present(name1) == true) break;
+                        if(local_network.is_present(name1) == true) break;
 
-                    else {
-                        system("CLS");
-                        std::cout<<"El router no existe"<<std::endl;
+                        else {
+                            system("CLS");
+                            std::cout<<"El router no existe"<<std::endl;
+                        }
                     }
-                }
 
-                while(true){
-                    std::cout<<"Ingrese el router destino: ";
-                    std::cin>>name2;
+                    while(true){
+                        std::cout<<"Ingrese el router destino: ";
+                        std::cin>>name2;
 
-                    if(local_network.is_present(name2) == true) break;
+                        if(local_network.is_present(name2) == true) break;
 
-                    else {
+                        else {
+                            system("CLS");
+                            std::cout<<"El router no existe"<<std::endl;
+                        }
+                    }
+
+                    if(name1 != name2) break;
+
+                    else{
                         system("CLS");
-                        std::cout<<"El router no existe"<<std::endl;
+                        std::cout<<"Los routers deben ser distintos"<<std::endl;
                     }
                 }
 
@@ -215,62 +243,95 @@ int main()
             case 5:{ //Eliminar enlace
 
                 while(true){
-                    std::cout<<"Ingrese el router origen: ";
-                    std::cin>>name1;
+                    while(true){
+                        std::cout<<"Ingrese el router origen: ";
+                        std::cin>>name1;
 
-                    if(local_network.is_present(name1) == true) break;
+                        if(local_network.is_present(name1) == true) break;
 
-                    else {
+                        else {
+                            system("CLS");
+                            std::cout<<"El router no existe"<<std::endl;
+                        }
+                    }
+
+                    while(true){
+                        std::cout<<"Ingrese el router destino: ";
+                        std::cin>>name2;
+
+                        if(local_network.is_present(name2) == true) break;
+
+                        else {
+                            system("CLS");
+                            std::cout<<"El router no existe"<<std::endl;
+                        }
+                    }
+
+                    if(name1 != name2 && local_network.is_present(name1, name2)) break;
+
+                    else{
                         system("CLS");
-                        std::cout<<"El router no existe"<<std::endl;
+                        std::cout<<"Los routers deben ser distintos"<<std::endl;
                     }
                 }
 
-                while(true){
-                    std::cout<<"Ingrese el router destino: ";
-                    std::cin>>name2;
+                while(true){ //confirmation
 
-                    if(local_network.is_present(name2) == true) break;
+                    std::cout<<"Esta a punto de desenlazar "<<name1<<" y "<<name2<<" (Esto puede afectar su existencia en la red). Desea continuar? (Y/N): "; std::cin>>aux;
 
+                    if(aux == "Y" || aux == "y"){
+
+                        local_network.modify_link(name1, name2, -1);
+
+                        std::cout<<"Enlace borrado correctamente";
+                        Sleep(1000);
+
+                        break;
+                    }
+                    else if(aux == "N" || aux == "n") break;
                     else {
                         system("CLS");
-                        std::cout<<"El router no existe"<<std::endl;
+                        std::cout<<"Opcion no valida"<<std::endl;
                     }
                 }
-
-                //confirmation maybe
-
-                local_network.delete_link(name1, name2);
-
                 break;
             }
             case 6:{ //Editar enlace
 
                 while(true){
-                    std::cout<<"Ingrese el router origen: ";
-                    std::cin>>name1;
+                    while(true){
+                        std::cout<<"Ingrese el router origen: ";
+                        std::cin>>name1;
 
-                    if(local_network.is_present(name1) == true) break;
+                        if(local_network.is_present(name1) == true) break;
 
-                    else {
+                        else {
+                            system("CLS");
+                            std::cout<<"El router no existe"<<std::endl;
+                        }
+                    }
+
+                    while(true){
+                        std::cout<<"Ingrese el router destino: ";
+                        std::cin>>name2;
+
+                        if(local_network.is_present(name2) == true) break;
+
+                        else {
+                            system("CLS");
+                            std::cout<<"El router no existe"<<std::endl;
+                        }
+                    }
+
+                    if(name1 != name2) break;
+
+                    else{
                         system("CLS");
-                        std::cout<<"El router no existe"<<std::endl;
+                        std::cout<<"Los routers deben ser distintos"<<std::endl;
                     }
                 }
 
-                while(true){
-                    std::cout<<"Ingrese el router destino: ";
-                    std::cin>>name2;
-
-                    if(local_network.is_present(name2) == true) break;
-
-                    else {
-                        system("CLS");
-                        std::cout<<"El router no existe"<<std::endl;
-                    }
-                }
-
-                cost = get_cost(false);
+                cost = get_cost(true);
 
                 local_network.modify_link(name1, name2, cost);
 
@@ -298,11 +359,30 @@ int main()
             }
             case 10:{ //Crear nueva red
 
-                //ask for confirmation***
+                while(true){
 
-                local_network.empty_network();
+                    std::cout<<"Esta a punto de borrar toda la red actual. Desea continuar? (Y/N): "; std::cin>>name1;
 
-                gen = false;
+                    if(name1 == "Y" || name1 == "y"){
+
+                        local_network.empty_network();
+                        gen = false;
+
+                        std::cout<<"Red borrada correctamente";
+                        Sleep(1000);
+
+                        break;
+                    }
+                    else if(name1 == "N" || name1 == "n") break;
+                    else {
+                        system("CLS");
+                        std::cout<<"Opcion no valida"<<std::endl;
+                    }
+                }
+                break;
+            }
+            case 11:{ //Salir
+                loop1 = false;
                 break;
             }
             }
